+++ app/Http/Controllers/CategoryController.php
@@ -4,11 +4,13 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Category;
+use Illuminate\Support\Facades\Gate;
 
 class CategoryController extends Controller
 {
     public function index()
     {
+        Gate::authorize('browse_categories');
         $dataType = 'category';
         $data = Category::all();
 
@@ -20,6 +22,7 @@ class CategoryController extends Controller
 
     public function store(Request $request)
     {
+        Gate::authorize('create_categories');
         $category = Category::create($request->all());
 
         return redirect()->route('categories.index')->with('message', 'Created Successfully!');
@@ -27,6 +30,7 @@ class CategoryController extends Controller
 
     public function show(Category $category)
     {
+        Gate::authorize('read_categories');
         $dataType = 'category';
         $data = $category;
 
@@ -38,6 +42,7 @@ class CategoryController extends Controller
     
     public function edit(Category $category)
     {
+        Gate::authorize('update_categories');
         $dataType = 'category';
         $data = $category;
 
@@ -49,6 +54,7 @@ class CategoryController extends Controller
 
     public function create()
     {
+        Gate::authorize('create_categories');
         $dataType = 'category';
         $data = null;
         return view('admin.bread.add-edit',compact(
@@ -59,6 +65,7 @@ class CategoryController extends Controller
 
     public function update(Request $request,Category $category)
     {
+        Gate::authorize('update_categories');
 
         $category->update($request->all());
 
@@ -67,6 +74,7 @@ class CategoryController extends Controller
 
     public function destroy(Category $category)
     {
+        Gate::authorize('delete_categories');
         $category->delete();
 
         return redirect()->route('categories.index')->with('message', 'Deleted Successfully!');
@@ -75,6 +83,7 @@ class CategoryController extends Controller
 
 
     public function bulkDestroy(Request $request){
+        Gate::authorize('delete_categories');
 
         Category::destroy($request->ids);
         return redirect()->route('categories.index')->with('message', 'Deleted Successfully!');
